feat(ezistock): add AI service with multi-agent system and alpha layer
- Add Python-based AI service with gRPC server for agent orchestration
- Implement multi-agent system: idea generator, investment advisor, news analyst, technical analyst, and strategy builder
- Add alpha layer with backtesting, data, deployment, model, and regime detection capabilities
- Implement feedback engine for accuracy and bias monitoring
- Add LLM router with caching and cost tracking for model optimization
- Add research module for factor snapshots, sector deep dives, and PDF generation
- Integrate proto-based gRPC services for agent, alpha, and feedback domains
- Add comprehensive test suite with mock LLM support
- Expand backend with new domain handlers: agent, analyst, consensus, knowledge, and fund management
- Refactor market domain with specialized handlers for exchange rates, gold, macro data, price boards, and trading hours
- Update project structure and Kiro configuration for multi-service architecture
- Add Kiro hooks for documentation auto-update, code quality review, and i18n translation

+++ backend/internal/handler/portfolio.go
@@ -12,7 +12,7 @@ import (
 // HandlePortfolioSummary returns the full portfolio overview for the authenticated user.
 // GET /api/portfolio/summary
 func (h *Handlers) HandlePortfolioSummary(c *gin.Context) {
-	userID := int64(c.MustGet("claims").(*model.JWTClaims).UserID)
+	userID := c.MustGet("claims").(*model.JWTClaims).UserID
 	summary, err := h.PortfolioEngine.GetPortfolioSummary(c.Request.Context(), userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
@@ -45,7 +45,7 @@ func (h *Handlers) HandleDeleteAsset(c *gin.Context) {
 // HandleGetTransactions returns all transactions for the authenticated user.
 // GET /api/portfolio/transactions
 func (h *Handlers) HandleGetTransactions(c *gin.Context) {
-	userID := int64(c.MustGet("claims").(*model.JWTClaims).UserID)
+	userID := c.MustGet("claims").(*model.JWTClaims).UserID
 	txns, err := h.TransactionLedger.GetTransactionsByUser(c.Request.Context(), userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
@@ -57,7 +57,7 @@ func (h *Handlers) HandleGetTransactions(c *gin.Context) {
 // HandleRecordTransaction records a new transaction.
 // POST /api/portfolio/transactions
 func (h *Handlers) HandleRecordTransaction(c *gin.Context) {
-	userID := int64(c.MustGet("claims").(*model.JWTClaims).UserID)
+	userID := c.MustGet("claims").(*model.JWTClaims).UserID
 	var tx model.Transaction
 	if err := c.ShouldBindJSON(&tx); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -75,7 +75,7 @@ func (h *Handlers) HandleRecordTransaction(c *gin.Context) {
 // HandlePortfolioPerformance returns performance metrics for the authenticated user.
 // GET /api/portfolio/performance?start=2024-01-01&end=2024-12-31
 func (h *Handlers) HandlePortfolioPerformance(c *gin.Context) {
-	userID := int64(c.MustGet("claims").(*model.JWTClaims).UserID)
+	userID := c.MustGet("claims").(*model.JWTClaims).UserID
 
 	start, end := parseDateRange(c)
 	metrics, err := h.PerformanceEngine.GetPerformanceMetrics(c.Request.Context(), userID, start, end)
@@ -89,7 +89,7 @@ func (h *Handlers) HandlePortfolioPerformance(c *gin.Context) {
 // HandlePortfolioRisk returns risk metrics for the authenticated user.
 // GET /api/portfolio/risk
 func (h *Handlers) HandlePortfolioRisk(c *gin.Context) {
-	userID := int64(c.MustGet("claims").(*model.JWTClaims).UserID)
+	userID := c.MustGet("claims").(*model.JWTClaims).UserID
 	metrics, err := h.RiskService.ComputeRiskMetrics(c.Request.Context(), userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
